requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.1.0
numpy>=1.22.0
torch>=2.0.0
torchvision>=0.15.0
pydantic==2.4.2
gunicorn==21.2.0 